+++ news-service/src/main/java/ru/clevertec/nms/utils/cache/impl/LFUCache.java
@@ -1,7 +1,6 @@
 package ru.clevertec.nms.utils.cache.impl;
 
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 import ru.clevertec.nms.utils.cache.Cache;
 
 import java.util.HashMap;
@@ -17,7 +16,7 @@ import java.util.Map;
 public class LFUCache<T> implements Cache<T> {
 
     private final int capacity;
-    private final Map<Integer, Node<T>> elements;
+    private final Map<Long, Node<T>> elements;
     private final Node<T> head;
     private final Node<T> tail;
 
@@ -43,7 +42,7 @@ public class LFUCache<T> implements Cache<T> {
      * @return value - if a value with the specified ID exists or else null
      */
     @Override
-    public T get(int key) {
+    public T get(long key) {
         Node<T> current = elements.get(key);
         T value = null;
         if (current != null) {
@@ -60,7 +59,7 @@ public class LFUCache<T> implements Cache<T> {
      * @param value value to add
      */
     @Override
-    public void put(int key, T value) {
+    public void put(long key, T value) {
         if (elements.containsKey(key)) {
             Node<T> current = elements.get(key);
             current.value = value;
@@ -82,7 +81,7 @@ public class LFUCache<T> implements Cache<T> {
      * @param key value ID to delete
      */
     @Override
-    public void delete(int key) {
+    public void delete(long key) {
         Node<T> removed = elements.remove(key);
         if (removed != null) {
             remove(removed);
@@ -136,7 +135,7 @@ public class LFUCache<T> implements Cache<T> {
      * @param <T> type that node should work with
      */
     private static class Node<T> {
-        int key;
+        long key;
         T value;
         int count;
         Node<T> prev;
@@ -145,7 +144,7 @@ public class LFUCache<T> implements Cache<T> {
         public Node() {
         }
 
-        public Node(int key, T value) {
+        public Node(long key, T value) {
             this.key = key;
             this.value = value;
             this.count = 1;
